requirements.txt
Django>=4.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0 